requirements.txt
streamlit
pandas
matplotlib
langgraph
langchain
langchain-openai
python-dotenv
kaleido
plotly